requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0

# AI/Chatbot dependencies
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0

# PDF extraction (optional but recommended for extract_courses.py)
pdfplumber>=0.11.0
tabula-py>=2.9.0
